requirements.txt
# Document parsing
python-docx==1.1.2
pdfplumber==0.11.4

# Vector DB
qdrant-client==1.12.1

# LLM
google-generativeai==0.8.3

# API
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Utils
python-dotenv==1.0.1
tqdm==4.67.1
pydantic==2.10.4
